% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocBuildReporter-package.R
\docType{package}
\name{BiocBuildReporter}
\alias{BiocBuildReporter-package}
\alias{BiocBuildReporter}
\title{BiocBuildReporter:  Functions to process a bioconductor build report database}
\description{
The BiocBuildReporter package provides access to years of Bioconductor build
system data, representing a comprehensive record of package builds across:
- **Thousands of packages** in the Bioconductor ecosystem
- **Multiple R versions** spanning several years of development
- **Multiple platforms** including Linux, macOS, and Windows
- **Different build stages** (install, build, check)
}
\details{
Please see vignette("BiocBuildReporter") for details on usage and pre-built
query functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lshep/BiocBuildReporter.git}
  \item Report bugs at \url{https://github.com/lshep/BiocBuildReporter/issues}
}

}
\author{
\strong{Maintainer}: Lori Shepherd \email{lori.shepherd@roswellpark.org} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Authors:
\itemize{
  \item Sean Davis \email{seandavi@gmail.com}
}

}
