% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetTables.R
\name{get_bbs_table}
\alias{get_bbs_table}
\title{Retrieve a remotely read parquet file of Bioconductor build data}
\usage{
get_bbs_table(
  tblname = c("build_summary", "info", "propagation_status"),
  useLocal = TRUE,
  updateLocal = FALSE
)
}
\arguments{
\item{tblname}{A valid parquet file name. Currently available: build_summary,
info, or propagation_status.}

\item{useLocal}{A logical indicating if files should be read/downloaded
locally or read remotely.}

\item{updateLocal}{A logical indicating if using locally cached files, if the
files should be re-downloaded/updated. Theorically build reports are updated
daily so the local data may be out of date if not updated.}
}
\value{
returns a tibble.
}
\description{
Bioconductor Build results are at times cumbersome to work with and
interrogate. Parquet files of condensed reports are provided remotely. These
functions read the remote parquet files using the arrow package and return a
tibble. Once retrieved the table is stored in a custom environment for easy
quick reference either to use ad hocly or within pre-defined queries provided
by this package. If useLocal is chosen, the remote parquet files are
downloaded locally and cached using BiocFileCache. If already downloaded, the
local file is utilized unless updateLocal is also TRUE in which case the file
will be redownloaded to the local location before reading into the
session. If useLocal is FALSE, the files will be read remotely.
}
\details{
The get_bbs_table returns a tibble but is also saved in a custom
environment for quick easy subsequent reference. There are three tables
available and can be specified in get_bbs_table: build_summary, info, or
propagation_status. build_summary returns a tibble with the results of each
stage of the Bioconductor build process. info returns a tibble with
information regarding the package maintainer and git commit
status. propagation_status returns a tibble with data on if a package
propagated to the community.
}
\examples{
info <- get_bbs_table("info")
info

library(dplyr)
# all the package Lori Shepherd maintains
info |> filter(Maintainer == "Lori Shepherd") |> distinct(Package)


}
\seealso{
get_all_bbs_tables
}
\author{
Lori Shepherd
}
