% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetQueries.R
\name{list_all_branches}
\alias{list_all_branches}
\title{Get list of all Bioconductor git branches}
\usage{
list_all_branches(infoTbl = NULL)
}
\arguments{
\item{infoTbl}{downloaded paraquet info table retrieved from
get_bbs_table('info') that includes git_branch information.}
}
\value{
character() vector listing all options for "branch"
}
\description{
Get list of allBioconductor git branches
}
\details{
Get list of all Bioconductor git branches
}
\examples{
list_all_branches()

}
\seealso{
get_bbs_table
}
