% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetQueries.R
\name{list_all_builders}
\alias{list_all_builders}
\title{Get list of all Bioconductor builders}
\usage{
list_all_builders(summaryTbl = NULL)
}
\arguments{
\item{summaryTbl}{downloaded paraquet build_summary table retrieved from
get_bbs_table('build_summary') that includes builder (node)  information.}
}
\value{
character() vector listing all options for "builder"
}
\description{
Get list of allBioconductor builders
}
\details{
Get list of all Bioconductor builders
}
\examples{
list_all_builders()

}
\seealso{
get_bbs_table
}
